backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pandas>=2.0.0
datasets>=2.19.0
google-genai>=1.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
